Move the WebView component to its own package
