+++ WooCommerce/src/main/kotlin/com/woocommerce/android/ui/onboarding/payments/GetPaidScreen.kt
@@ -9,7 +9,7 @@ import androidx.compose.ui.res.stringResource
 import com.woocommerce.android.R.string
 import com.woocommerce.android.ui.common.wpcomwebview.WPComWebViewAuthenticator
 import com.woocommerce.android.ui.compose.component.Toolbar
-import com.woocommerce.android.ui.compose.component.WCWebView
+import com.woocommerce.android.ui.compose.component.web.WCWebView
 import org.wordpress.android.fluxc.network.UserAgent
 
 @Composable
